Matplotlib is now in the pipeline
As matplotlib is a general package for plotting and it is widely
used in science, we have added it to the pipeline.

When installing a dependency of matplotlib `python-dateutil', we
found a conflict in the download of the tarball. This is because
the name has a dash (-) in the middle. In addition, the name starts
with 'python', so it is the same as the python itself. Now it is
possible to install any package with any name, just adding an elif
in before the URL direction.