Git hooks for metastore check for the existance of metastore
Until now, once the Git hooks have been installed (after the
installation of Metastore), if metastore doesn't exist (for example by
manually deleting the build directory for a re-build with same
configurations as before) we can't run `git commit' and `git checkout'
will print an ugly warning.

With this commit, the two Git hooks check for the existance of Metastore
and if it doesn't exist, they won't do anything.