Python installation and related packages have been split
Until now the installation of Python and its packages (numpy, astropy,
astroquery, etc.) were done in the same `makefile'.

With this commit the installation of Python and its packages have been
split and now it is independent of the other programs. The installation
of all Python packages needs to be written explicitely because pip is
not used anymore.