GNU GCC is now in the pipeline
Until now, the pipeline was not installing its own `gcc' but using the
system one by making a symbolic link.

With this commit, GNU GCC has been added into the pipeline. Right now
the installation does not work on Mac OS system beause of some conflicts
with `clang', but in principle it should work on GNU Linux distributions.