for-group gives write permission to all built software in the end
Since the `install' script also sets permissions manually, the permissions
that we define in `for-group' don't usually affect the installed
files. Therefore the installed files of one user can't be modified/deleted
by another. With this commit, after for-group finishes configuration, it
also adds the write flag for all group members in the whole installation
directory.