.gitignore with copyright using better search for copyright notice
In the previous commit, a copyright notice was added after a systematic
search of the version controlled files. However, we missed `.gitignore'
(because we were discarding those with the `*.git*' pattern to avoid files
in the `.git' directory). This has been fixed by using this command (in the
top project directory) instead:

    for f in $(find ./ -type f); do \
      if [[ $f != ./.git/* ]]; then \
        n=$(grep -i copyright $f | wc -l); \
        echo "$n $f"; \
      fi; \
    done | awk '$1==0'