Python packages are installed as high level program dependencies
Until this commit, the installation of all Python packages were
done in a separate Makefile.

With this commit, the pipeline install Python packages as part of the
hight level software. All Python packages rules  them remain in a
separate Makefile, but this Makefile is included in the high level
dependency `reproduce/src/make/dependencies.mk'.