Configure script also accepts short options with no delimiter
Until now, the short options to the configure script needed a delimiter
(either white-space or an `=') between the name and value.

With this commit, for short options, it also accepts the value immediately
touching the option name.

Also, when trying to fine the absolute address of a given path, a check was
added to abort if it doesn't exist.