Corrected typo in GCC build recipe
For the tests, we had just set an absurd value for a test in the GCC build
recipe to always fail, but we had forgot to fix it. It is now corrected.

Also the order of making `g++' and `gfortran' was reversed for easier
readablility (it doesn't matter which one is done first, it only matters
that `gcc' be done last).