Using gnuastro name for Gnuastro's target, not astnoisechisel
Until recently we were using an actual installed executable file for the
programs. So for Gnuastro, the target was called `astnoisechisel'. But
recently, this approach was changed and the target for each software is a
simple text file with the official software name and version.

So with this commit, we are simply using `gnuastro' for Gnuastro, not
`astnoisechisel'.