Corrections in configure script and astroquery, astropy rules
The step where we check the possibility of using `sys/cdefs.h' was still
using `$$' for shell variables (in Make), not `$' (for the shell). This was
corrected.

Also, since Astropy needs two citations, the `,' in the citation command
would conflict with Make's parsing. So we just used an `echo' command to
re-write the version info.

In Astroquery, the prerequisite list was just reordered by length to be
more clear to the eye.