Using Clang on Mac OS systems for pkg-config
Yahya Sefidbakht reported the following error when building Pkg-config on
his Mac OS system (using GCC, not Clang). It is apparently because his
version of GCC doesn't support some speical feature on Mac that is
necessary to build Glib as part of Pkg-config.

With this commit, on Mac systems, for pkg-config we are explicity asking to
build with Clang (through the `CC' flag).