for-group: Options passed to configure or make
Until now, the `for-group' script could only two options to Make and didn't
pass anything to `./configure'. With this commit, any thing given after the
first two tokens to `for-group' is passed directly to `./configure' or
Make.

Also, the test for the existance of the group has been modified to avoid
making a temporary file.