Using bin executable in patchelf for awk and bash
Until this commit, we were using the target (version number of the
program) in the `patchelf' for `awk' and `bash'. This makes an incorrect
linking in libraries because the target is not the bin program but just
a plain text containing the version number of the program.

With this commit we fix this issue by setting in the patchelf of `awk'
and `bash' the bin executable, and not the target (version number).