Unzip set as prerequisite of Numpy and Setuptools
Since we mixed the installation of Python packages with all other
software, it may occur that some Python packages start to be installed
before having installed `unzip'. As a consecuence, they could not be
decompressed and the installation will fail. In particular, tarballs of
Numpy and Setuptools are .zip files.

With this commit, we fix this issue by setting `unzip' as a prerequisit
of Numpy and Setuptools.