Configure script dealing properly with empty software directories
Until now, we hadn't actually tested the case where a whole software
directory (Python modules in particular) is empty. So the configure script
finished with some errors in this case.

With this commit, this step of the configure script was modified to deal
with such cases cleanly.

Also, in `initialize.mk', I added a `-f' to the symbolic link command, so
it doesn't complain if the file link already exists.