Acknowledged Scipy-related packages: Cython, Matplotlib, Numpy and Scipy
Until now, name and version of all Python packages were indicated in the
final paper, but not the main paper of them (if it exists).

With this commit, some Python packages (Cython, Matplotlib, Numpy and
Scipy) are now properly acknoledged by citating the source paper.
`mpi4py' is also cited although this package is not yet included into
the pipeline.