File is built as a dependency of GCC
Until now, we did not have `file'.  It was in other project, where a
problem with `Astrometry-net' software, ends up with the necessity of
having `file' into the pipeline.

With this commit, we add `file' to the project. Since it is a low level
program, it is set in `dependencies-basic.mk' as a prerequisite of GCC.