Minor typo corrections in previous commits
Until now, even though `file' was a dependency of `gcc', it was still
listed as a `top-level-programs'. Also, we weren't including the new
`tex/dependencies' in the distribution tarball (with `make dist').

With this commit, both issues are solved and also, as a cosmetic change,
the GCC prerequisites of the same line-length were ordered alphabetically.