Cairo library and its dependency Pixman, are now included
In this commit we add Cairo library. The motivation of including this
library is because it is a dependency of Astrometry-net (which is a
common software used in astronomy projects).  In principle it will not
be build by default by de pipeline, but we will keep the rule just in
case it is needed.

Pixman is another library which is a prerequisite of Cairo, it is also
built with this commit.