Netpbm and its dependencies are now included
With this commit, the installation of `netpbm' has been included. This
software has a crazy dialogue installation that we override by giving to
it the answers (they differs on GNU/Linux and Mac OS systems). `libxml2'
also has been built (a remaining dependency of `netpbm').

In this commit, `libpng' version has been downgraded because it is
required by `netpbm'.

The motivation of installing `netpbm' is that it is a dependency of
Astrometry-net software.