Binutils is built on GNU/Linux as GCC dependency, not in MAC OS
Until this commit, we didn't  install Binutils. However, we need `strip'
for installing Netpbm, and `strip' is part of the Binutil software.

With this commit, we include Binutils as a dependency of GCC for
GNU/Linux systems. For Mac OS systems  we create a symbolic link to
`strip' just after the generation of the symbolic link to `gcc'.