SWarp is now into the pipeline
SWarp resamples and co-adds together FITS images using any arbitrary
astrometric projection defined in the WCS standard. It is a very useful
program for astronomy and that is why we added it to the pipeline.

With this commit, we also cite the paper for Astrometry-net that until
now was missing.