Corrected bibtex entry for Astrometry-net and Swarp
Until now, there were erros in the citation of Astrometry-net and Scamp
papers.

With this commit, we fix these problems. The Swarp bibtex has also been
modify to follow the stetic of the citation style we have right now in
the project.

We also added the `dependency-bib.tex' as a prerequisite of `paper.bbl'.