Removed ATLAS as a prerequisite of Scamp
Until this commit, ATLAS was a prerequisite of Scamp. The documentation
says that. However, we have been able to install Scamp without having
ATLAS installed. As a consecuence, ATLAS does not have to be a
prerequisite of Scamp anymore.

With this commit, we remove ATLAS as prerequisite of Scamp. We also put
available software in the template as optional because not all people
are going to use these software.