Awk, coreutils and sed are now prerequisites of metastore
Until now, `metastore' did not depend on the necessaries programs that
we use to install it (`awk', `coreutils' and `sed'). They are not
official dependencies of `metastore', but we need them to install it.

With this commit, we put these programs as prerequisites of `metastore'
in order to be able to install it without any problem.