Python package SymPy added into the template
With this commit, we add `sympy' Python package into the template. This
is a package to do symbolic mathematics.

The motivation is that it is a prerequisite of `sip_tpv' Python package,
which is useful to convert SIP distorsion coefficients into PV
coefficients (in the context of astronomical images). However, the
availability of `sympy' in the template  will be useful for anyone
interested in this package.