Added --enable-single option in FFTW installation
Until this commit, FFTW was installed by default, without the single
precission option. However, when installing `sextractor' in a new system
we found an error in the installation complaining about not having
single precission files of FFTW. More explicity the error was: "FFTW
single precision library files not found in [...]".

With this commit, we fix this problem by passing the option
`--enable-single' in the installation of FFTW.