Lack of semicolons in Binutils installation rule for Mac OS fixed
Until this commit, in Mac OS we were installing `binutils' as symbolic
links of various programs (`as', `ar', `ld', `nm',` ps' and  `ranlib').
However, there was a missing semicolon at the end of each line. As a
consecuence, the installation of `binutils' on Mac OS systems failed.

With this commit, we fix this problem by typing a semicolon at the end
of each makelink line.