Libpng updated (version 1.6.37)
Until this commit, we were using `libpng' version 1.4.22 (out dated).
The reason is that  with `libpng` versions > 1.5, there was some
problems with `netpbm'.  But we have update `netpbm' version and now
there is no conflict betweem them.

With this commit, we increase the version of `libpng' to have the most
recent one.