Astrometry-net installation modified to avoid two reporting steps
Until this commit, we had a problem in the installation of
Astrometry-net in Mac OS systems. During the runing of the Makefile, it
stopped at the reporting stage because it tried to do a couple of steps
that in Mac OS are not possible:

$ cat /proc/cpuinfo
$ free

The reason is that `/proc/cpuinfo' and `free' doesn't exist in Mac OS.

With this commit, we override these two steps. Since they are only for
reporting, they are not mandatory. So, we avoid doing them by using
`sed' before running the Makefile.