Added numpy and six as prerequisites of h5py
Until now, we were missing `numpy' and `six' as prerequisites of `h5py'.
Because we did the configure step with all cores, `numpy' and `six'
allways were built before `h5py' with any crash.  However, in a single
thread configure run we noticed that we were missing these two
dependencies.

With this commit, we fix this issue by setting `numpy' and `six' as
prerequisites of `h5py'.