Running PatchELF on all programs after Coreutils
Even though the Coreutils configure script says that it adds `rpath' to its
executables, its executables don't have it. As a result, it goes to use the
host operating system libraries, causing failures when they don't match
with the template.

Unfortunately after some tests with the configure script, I couldn't find
any way to include `rpath'! Even though `-rpath-link' is present in all
linking commands during Coreutils' build, the installed files still don't
have it!

So I was forced to use PatchELF. However, one problem is that Coreutils
installs many programs, not just one like Bash or AWK.

As a brute-force solution, with this commit, we are running PatchELF on all
the installed programs. It won't hurt those that already have it, but it
will fix all those that don't.

With this commit, I am also making the following non-related small changes:

 - In `build-rules.mk', the `&&' characters were placed at the start of the
   line for better readability.

 - Bash, Readline and NCURSES are now built in parallel on systems that
   support it (during `basic.mk').